requirements.txt
numpy~=1.16.2
Pillow~=7.1.2
xgboost~=1.1.1
seaborn~=0.10.1
pandas~=1.0.4

